Subsume `helpers.facilities` in `lib.facilities`
These things did not belong in the `helpers` due to their (formal and
logical) dependency on `Building`.
