+++ pycroft/lib/facilities.py
@@ -3,6 +3,7 @@ pycroft.lib.facilities
 ~~~~~~~~~~~~~~~~~~~~~~
 """
 import logging
+import re
 import typing as t
 from collections import defaultdict
 from dataclasses import dataclass
@@ -203,3 +204,30 @@ def suggest_room_address_data(building: Building) -> RoomAddressSuggestion | Non
                        row_to_suggestion(rest[0]), rest[0].count)
 
     return suggestion
+
+
+def sort_buildings(buildings: t.Iterable[Building]) -> list[Building]:
+    def make_sort_key(building: Building) -> tuple[str, str | tuple[int, str]]:
+        s = re.split(r"(\d+)([a-zA-Z]?)", building.number)
+        if len(s) != 4:
+            return building.street, building.number  # split unsuccessful
+        return building.street, (int(s[1]), s[2].lower())
+
+    return sorted(buildings, key=make_sort_key)
+
+
+def determine_building(shortname: str | None = None, id: int | None = None) -> Building:
+    """Determine building from shortname or id in this order.
+
+    :param shortname: The short name of the building
+    :param id: The id of the building
+
+    :return: The unique building
+    """
+    if shortname:
+        return t.cast(
+            Building, Building.q.filter(Building.short_name == shortname).one()
+        )
+    if id:
+        return Building.get(id)
+    raise ValueError("Either shortname or id must be given to identify the building!")
